Cleanup the examples
