+++ examples/pbc/03-cplxCASCI.py
